# packages/contracts/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations/{organizationAlias}/tournaments/{tournamentAlias}/stages": {
      "post": {
        "description": "Number, name and format all default. Fixtures are not generated here — publish a seed order via POST .../stages/:stageNumber/seeding afterward.",
        "operationId": "StagesController_create",
        "parameters": [
          {
            "name": "organizationAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournamentAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a stage from the tournament’s accepted registrations",
        "tags": [
          "stages"
        ]
      }
    },
...
  },
  "info": {
    "title": "CopaLibre API",
    "description": "Self-hosted tournament management API.\n\nAuthentication is JWT Bearer, validated against the operator-selected\nidentity provider’s JWKS. Browsers use Authorization Code + PKCE and\nhold the access token in memory only — never localStorage. A token\npassed as a query parameter is ignored, because URLs leak into proxy\nlogs, browser history, metrics, traces and screenshots.\n\nEvery route declares one security plane: public-read,\nauthenticated-interaction, admin-control, or integration.\nauthenticated-interaction and admin-control share this same bearer\ntransport and differ only in authorization policy.",
    "version": "4.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Short-lived access token. Strict stateless mode is the default: no refresh credential is persisted, so a reload reauthenticates."
      }
    },
    "schemas": {
...
      "ProblemResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 403
          },
          "message": {
            "type": "string",
            "example": "subject may only act on their own records"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CreateStageRequest": {
        "type": "object",
        "properties": {
          "number": {
            "type": "number",
            "description": "Defaults to the tournament’s next sequential stage number. Refused as a conflict if a stage with this number already exists.",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Defaults to \"Stage {number}\".",
            "example": "Fase de grupos"
          },
          "format": {
            "type": "string",
            "description": "Defaults to the tournament’s own configured format. Validated against the tournament’s discipline descriptor when supplied.",
            "example": "round-robin"
          }
        }
      },
      "StageResponse": {
        "type": "object",
        "properties": {
          "stageId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid",
            "description": "The tournament edition this stage belongs to"
          },
          "number": {
            "type": "number",
            "description": "1-based sequential number within the tournament",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Fase de grupos"
          },
          "format": {
            "type": "string",
            "example": "round-robin"
          }
        },
        "required": [
          "stageId",
          "seasonId",
          "number",
          "name",
          "format"
        ]
      },
...
    }
  }
}